refactor: task.goに切り出す
